Move CategoryRepositories to persistence.category package
